ATHOS-129 feat(Dashboard/custo): Adiciona modelos para dashboard custo
