--- apps/dashboards/custo/views.py
@@ -1,7 +1,0 @@
-from django.http import JsonResponse
-from django.views.decorators.http import require_http_methods
-
-
-@require_http_methods(["GET"])
-def index(request):
-    return JsonResponse({'resposta': 'Olá mundo'})
